writedefault 検知機能追加
