+++ Assets/HhotateA/AvatarModifyTool/Editor/WindowBase.cs
@@ -8,6 +8,7 @@ This software is released under the MIT License.
 http://opensource.org/licenses/mit-license.php
 */
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor;
 using Object = UnityEngine.Object;
@@ -249,6 +250,35 @@ namespace HhotateA.AvatarModifyTools.Core
             EditorGUILayout.LabelField( "Version " + EnvironmentVariable.version ,signature);
             EditorGUILayout.LabelField( EnvironmentVariable.githubLink ,signature);
             // EditorGUILayout.LabelField( EnvironmentVariable.githubLink ,signature);
+        }
+
+        public void DetectAnimatorError()
+        {
+            if (!writeDefault)
+            {
+                var writeDefaultLayers = FindWriteDefault();
+                if (writeDefaultLayers.Count > 0)
+                {
+                    status.Warning("Complete Setup (Detect WriteDefault Layers)");
+                    foreach (var c in writeDefaultLayers)
+                    {
+                        Debug.LogWarning("Detect WriteDefault Layer : " + c);
+                    }
+                }
+            }
+        }
+        
+        List<string> FindWriteDefault()
+        {
+#if VRC_SDK_VRCSDK3
+            AvatarModifyTool mod = new AvatarModifyTool(avatar);
+            ApplySettings(mod);
+
+            var writeDefaultLayers = mod.HasWriteDefaultLayers();
+            return writeDefaultLayers;
+#else
+            return new List<string>();
+#endif
         }
 
         public class StatusView
